+++ dotnet/Services/VtexAPIService.cs
@@ -808,8 +808,10 @@ namespace ShipStation.Services
 
         public async Task<string> CheckCancelledOrders(DateTime date)
         {
+            bool updated = false;
+            int daysToCheck = 2;
             StringBuilder sb = new StringBuilder();
-            ListOrdersResponse listOrdersResponse = await _shipStationAPIService.ListOrders($"modifyDateStart={date.AddDays(-2)}&modifyDateEnd={date}&orderStatus={ShipStationConstants.ShipStationOrderStatus.Canceled}&pageSize=500");
+            ListOrdersResponse listOrdersResponse = await _shipStationAPIService.ListOrders($"modifyDateStart={date.AddDays(-daysToCheck)}&modifyDateEnd={date}&orderStatus={ShipStationConstants.ShipStationOrderStatus.Canceled}&pageSize=500&sortBy=ModifyDate&sortDir=DESC");
             Console.WriteLine($"CheckCancelledOrders pages={listOrdersResponse.Pages}");
             if (listOrdersResponse != null)
             {
@@ -824,9 +826,78 @@ namespace ShipStation.Services
                         string orderState = vtexOrder.Status;
                         Console.WriteLine($"{orderId} {orderState}");
                         sb.AppendLine($"{orderId} {orderState}");
-                        if (!orderState.Equals(ShipStationConstants.VtexOrderStatus.Cancelled))
+                        if (!orderState.Equals(ShipStationConstants.VtexOrderStatus.Canceled))
                         {
-                            bool updated = await SetOrderStatus(orderId, ShipStationConstants.VtexOrderStatus.Cancel);
+                            long totalItemQuantitySS = order.Items.Sum(i => i.Quantity);
+                            long totalItemQuantityVtex = vtexOrder.Items.Sum(i => i.Quantity);
+                            if (totalItemQuantitySS == totalItemQuantityVtex)
+                            {
+                                updated = await SetOrderStatus(orderId, ShipStationConstants.VtexOrderStatus.Cancel);
+                                _context.Vtex.Logger.Info("CheckCancelledOrders", "CancelOrder", $"Order {order} cancelled? {updated}");
+                            }
+                            else
+                            {
+                                ChangeOrderRequest changeOrderRequest = new ChangeOrderRequest
+                                {
+                                    OrderId = orderId,
+                                    Reason = ShipStationConstants.ORDER_CHANGE_REASON,
+                                    ItemsRemoved = new List<ChangeItem>(),
+                                    RequestId = order.ModifyDate
+                                };
+
+                                if (vtexOrder.ChangesAttachment != null && vtexOrder.ChangesAttachment.ChangesData != null)
+                                {
+                                    var itemsRemovedVtex = vtexOrder.ChangesAttachment.ChangesData.SelectMany(c => c.ItemsRemoved);
+                                    ListOrdersResponse listOrdersByOrderNumberResponse = await _shipStationAPIService.ListOrders($"orderNumber={orderId}");
+                                    var cancelledItemsShipstation = listOrdersByOrderNumberResponse.Orders.Where(o => o.OrderStatus.Equals(ShipStationConstants.ShipStationOrderStatus.Canceled)).SelectMany(i => i.Items);
+                                    var notCancelledItemsShipstation = listOrdersByOrderNumberResponse.Orders.Where(o => !o.OrderStatus.Equals(ShipStationConstants.ShipStationOrderStatus.Canceled)).SelectMany(i => i.Items);
+
+                                    foreach (OrderItem cancelledItem in order.Items)
+                                    {
+                                        long totalCancelledForSku = cancelledItemsShipstation.Where(o => o.Sku.Equals(cancelledItem.Sku)).Sum(i => i.Quantity);
+                                        long qntyAlreadyRemovedForSku = itemsRemovedVtex.Where(o => o.Id.Equals(cancelledItem.Sku)).Sum(i => i.Quantity);
+                                        Console.WriteLine($"[-] | [-] | [-] | [-] | [-]     ({cancelledItem.Sku}): {qntyAlreadyRemovedForSku} < {totalCancelledForSku}");
+                                        if (qntyAlreadyRemovedForSku < totalCancelledForSku)
+                                        {
+
+                                            ChangeItem changeItem = new ChangeItem
+                                            {
+                                                Id = cancelledItem.LineItemKey,
+                                                Price = (long)(cancelledItem.UnitPrice * 100) * cancelledItem.Quantity,
+                                                Quantity = cancelledItem.Quantity
+                                            };
+
+                                            changeOrderRequest.ItemsRemoved.Add(changeItem);
+                                        }
+                                    }
+                                }
+                                else
+                                {
+                                    foreach (OrderItem cancelledItem in order.Items)
+                                    {
+                                        ChangeItem changeItem = new ChangeItem
+                                        {
+                                            Id = cancelledItem.LineItemKey,
+                                            Price = (long)(cancelledItem.UnitPrice * 100) * cancelledItem.Quantity,
+                                            Quantity = cancelledItem.Quantity
+                                        };
+
+                                        changeOrderRequest.ItemsRemoved.Add(changeItem);
+                                    }
+                                }
+
+                                if (changeOrderRequest.ItemsRemoved.Count > 0)
+                                {
+                                    ChangeOrderResponse changeOrderResponse = await ChangeOrder(orderId, changeOrderRequest);
+                                    updated = changeOrderResponse != null;
+                                    _context.Vtex.Logger.Info("CheckCancelledOrders", "ChangeOrder", $"ChangeOrderRequest: {JsonConvert.SerializeObject(changeOrderRequest)}  ChangeOrderResponse: {JsonConvert.SerializeObject(changeOrderResponse)}");
+                                }
+                                else
+                                {
+                                    Console.WriteLine("CheckCancelledOrders - Skipping...");
+                                }
+                            }
+
                             sb.AppendLine($"Order {orderId} updated? {updated}");
                         }
                     }
@@ -839,6 +910,42 @@ namespace ShipStation.Services
             }
 
             return sb.ToString();
+        }
+
+        public async Task<ChangeOrderResponse> ChangeOrder(string orderId, ChangeOrderRequest changeOrderRequest)
+        {
+            ChangeOrderResponse changeOrderResponse = null;
+            string jsonSerializedRequest = JsonConvert.SerializeObject(changeOrderRequest);
+            var request = new HttpRequestMessage
+            {
+                Method = HttpMethod.Post,
+                RequestUri = new Uri($"http://{this._httpContextAccessor.HttpContext.Request.Headers[ShipStationConstants.VTEX_ACCOUNT_HEADER_NAME]}.vtexcommercestable.com.br/api/oms/pvt/orders/{orderId}/changes"),
+                Content = new StringContent(jsonSerializedRequest, Encoding.UTF8, ShipStationConstants.APPLICATION_JSON)
+            };
+
+            request.Headers.Add(ShipStationConstants.USE_HTTPS_HEADER_NAME, "true");
+            string authToken = this._httpContextAccessor.HttpContext.Request.Headers[ShipStationConstants.HEADER_VTEX_CREDENTIAL];
+            if (authToken != null)
+            {
+                request.Headers.Add(ShipStationConstants.AUTHORIZATION_HEADER_NAME, authToken);
+                request.Headers.Add(ShipStationConstants.VTEX_ID_HEADER_NAME, authToken);
+                request.Headers.Add(ShipStationConstants.PROXY_AUTHORIZATION_HEADER_NAME, authToken);
+            }
+
+            //MerchantSettings merchantSettings = await _shipStationRepository.GetMerchantSettings();
+            //request.Headers.Add(ShipStationConstants.APP_KEY, merchantSettings.AppKey);
+            //request.Headers.Add(ShipStationConstants.APP_TOKEN, merchantSettings.AppToken);
+
+            var client = _clientFactory.CreateClient();
+            var response = await client.SendAsync(request);
+            string responseContent = await response.Content.ReadAsStringAsync();
+            Console.WriteLine($"ChangeOrder for Order '{orderId}' [{response.StatusCode}] {responseContent}");
+            if(response.IsSuccessStatusCode)
+            {
+                changeOrderResponse = JsonConvert.DeserializeObject<ChangeOrderResponse>(responseContent);
+            }
+
+            return changeOrderResponse;
         }
 
         private long ToCents(double asDollars)